* follow type.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@15584 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ include/ruby/intern.h
@@ -515,7 +515,7 @@ VALUE rb_str_unlocktmp(VALUE);
 VALUE rb_str_dup_frozen(VALUE);
 VALUE rb_str_plus(VALUE, VALUE);
 VALUE rb_str_times(VALUE, VALUE);
-int rb_str_sublen(VALUE, int);
+long rb_str_sublen(VALUE, long);
 VALUE rb_str_substr(VALUE, long, long);
 VALUE rb_str_subseq(VALUE, long, long);
 void rb_str_modify(VALUE);
